+++ LearnsMate/src/main/java/intbyte4/learnsmate/lecture/service/LectureServiceImpl.java
@@ -9,11 +9,9 @@ import intbyte4.learnsmate.lecture.domain.dto.TutorLectureVideoCountDTO;
 import intbyte4.learnsmate.lecture.domain.entity.Lecture;
 import intbyte4.learnsmate.lecture.mapper.LectureMapper;
 import intbyte4.learnsmate.lecture.repository.LectureRepository;
-import intbyte4.learnsmate.member.domain.dto.MemberDTO;
-import intbyte4.learnsmate.member.domain.entity.Member;
 import intbyte4.learnsmate.member.service.MemberService;
-import intbyte4.learnsmate.videobylecture.domain.dto.CountVideoByLectureDTO;
-import intbyte4.learnsmate.videobylecture.service.VideoByLectureService;
+import intbyte4.learnsmate.video_by_lecture.domain.dto.CountVideoByLectureDTO;
+import intbyte4.learnsmate.video_by_lecture.service.VideoByLectureService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
