+++ LearnsMate/src/main/java/intbyte4/learnsmate/lecture_video_by_student/domain/dto/LectureVideoByStudentDTO.java
@@ -1,0 +1,14 @@
+package intbyte4.learnsmate.lecture_video_by_student.domain.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public class LectureVideoByStudentDTO {
+    private Long videoCode;
+    private Long lectureByStudentCode;
+    private Boolean lectureStatus;
+}
