--- LearnsMate/src/main/java/intbyte4/learnsmate/lecturevideobystudent/domain/dto/LectureVideoByStudentDTO.java
@@ -1,14 +1,0 @@
-package intbyte4.learnsmate.lecturevideobystudent.domain.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-public class LectureVideoByStudentDTO {
-    private Long videoCode;
-    private Long lectureByStudentCode;
-    private Boolean lectureStatus;
-}
