--- LearnsMate/src/main/java/intbyte4/learnsmate/lecturevideobystudent/service/LectureVideoByStudentService.java
@@ -1,9 +1,0 @@
-package intbyte4.learnsmate.lecturevideobystudent.service;
-
-import intbyte4.learnsmate.lecturevideobystudent.domain.dto.LectureVideoProgressDTO;
-
-import java.util.List;
-
-public interface LectureVideoByStudentService {
-    List<LectureVideoProgressDTO> getVideoProgressByStudent(Long studentCode);
-}
