+++ LearnsMate/src/main/java/intbyte4/learnsmate/video_by_lecture/controller/VideoByLectureController.java
@@ -1,0 +1,7 @@
+package intbyte4.learnsmate.video_by_lecture.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class VideoByLectureController {
+}
