+++ LearnsMate/src/main/java/intbyte4/learnsmate/video_by_lecture/domain/dto/CountVideoByLectureDTO.java
@@ -1,0 +1,15 @@
+package intbyte4.learnsmate.video_by_lecture.domain.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class CountVideoByLectureDTO {
+    private Long lectureCode;
+    private long videoCount;
+}
