+++ LearnsMate/src/main/java/intbyte4/learnsmate/video_by_lecture/domain/dto/VideoByLectureDTO.java
@@ -1,0 +1,18 @@
+package intbyte4.learnsmate.video_by_lecture.domain.dto;
+
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public class VideoByLectureDTO {
+
+    private Long videoCode;
+    private String videoLink;
+    private String videoTitle;
+    private Long lectureCode;
+
+}
