+++ LearnsMate/src/main/java/intbyte4/learnsmate/video_by_lecture/service/VideoByLectureServiceImpl.java
@@ -1,0 +1,21 @@
+package intbyte4.learnsmate.video_by_lecture.service;
+
+import intbyte4.learnsmate.video_by_lecture.domain.dto.CountVideoByLectureDTO;
+import intbyte4.learnsmate.video_by_lecture.repository.VideoByLectureRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class VideoByLectureServiceImpl implements VideoByLectureService {
+
+    private final VideoByLectureRepository videoByLectureRepository;
+
+    // 강의의 동영상 개수 조회
+    @Override
+    public CountVideoByLectureDTO getVideoByLecture(Long lectureCode) {
+        long videoCount = videoByLectureRepository.countByLectureCode(lectureCode);
+        return new CountVideoByLectureDTO(lectureCode, videoCount);
+    }
+
+}
