--- LearnsMate/src/main/java/intbyte4/learnsmate/videobylecture/controller/VideoByLectureController.java
@@ -1,7 +1,0 @@
-package intbyte4.learnsmate.videobylecture.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class VideoByLectureController {
-}
