--- LearnsMate/src/main/java/intbyte4/learnsmate/videobylecture/service/VideoByLectureService.java
@@ -1,8 +1,0 @@
-package intbyte4.learnsmate.videobylecture.service;
-
-import intbyte4.learnsmate.videobylecture.domain.dto.CountVideoByLectureDTO;
-
-public interface VideoByLectureService {
-    // 강의의 동영상 개수 조회
-    CountVideoByLectureDTO getVideoByLecture(Long lectureCode);
-}
